backend/requirements.txt
Django==4.2
youtube-transcript-api==0.6.1
requests==2.31.0
fpdf2==2.7.4